% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-methods.R
\docType{methods}
\name{as.list,configuration-method}
\alias{as.list,configuration-method}
\alias{as.list}
\title{as.list}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{configuration object}

\item{...}{other arguments (not used, but included for consistency with base
generic)}
}
\value{
a named list
}
\description{
Coerce configuration object to list
}
\examples{
\dontrun{
as.list(configuration_object)
}

}
