% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-methods.R
\docType{methods}
\name{as.list,minid-method}
\alias{as.list,minid-method}
\alias{as.list}
\title{as.list}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{minid object}

\item{...}{other arguments (not used, but included for consistency with base
generic)}
}
\value{
a named list
}
\description{
Coerce minid object to list
}
\examples{
\dontrun{
as.list(minid_object)
}

}
