% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{checksum<-}
\alias{checksum<-}
\title{checksum}
\usage{
checksum(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new checksum string}
}
\value{
The minid object with specified checksum string
}
\description{
Set the minid checksum
}
\examples{
\dontrun{
a_hash <-"6484968f81afac84857d02b573b0d589fb2f9582a2b920572830dc5781e0a53c"
checksum(minid_object) <- a_hash
}

}
