% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{checksum}
\alias{checksum}
\alias{checksum,minid-method}
\title{checksum}
\usage{
\S4method{checksum}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid checksum (as character string)
}
\description{
Get the minid checksum
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the checksum slot
}}

\examples{
data("demo2", package = "minidtools")
checksum(demo2)

}
