% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{checksum_function<-}
\alias{checksum_function<-}
\title{checksum_function}
\usage{
checksum_function(x) <- "checksum_function string"
}
\arguments{
\item{x}{minid object}
}
\value{
The minid object with specified checksum_function string
}
\description{
Set the minid checksum_function
}
\examples{
\dontrun{
checksum_function(minid_object) <- "SHA"
}

}
