% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{checksum_function}
\alias{checksum_function}
\alias{checksum_function,minid-method}
\title{checksum_function}
\usage{
\S4method{checksum_function}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid checksum_function (as character string)
}
\description{
Get the minid checksum hashing function
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the checksum_function slot
}}

\examples{
data("demo", package = "minidtools")
checksum_function(demo)

}
