% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{code<-}
\alias{code<-}
\title{code}
\usage{
code(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new user access code string}
}
\value{
The configuration object with specified user access code
}
\description{
Set the minid user's access code
}
\examples{
\dontrun{
code(configuration_object) <- "XXXX-XXX-XXXXX-XXXX-XXXX"
}

}
