% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/configuration-getter-methods.R, R/configuration-setter-methods.R
\docType{methods}
\name{code}
\alias{code}
\alias{code,configuration-method}
\alias{code<-,configuration-method}
\title{code}
\usage{
code(x)

\S4method{code}{configuration}(x)

\S4method{code}{configuration}(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new code string}
}
\value{
The code
}
\description{
Get the code from the configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{configuration}: getter method for the code slot

\item \code{configuration}: setter method for the code slot
}}

\examples{
\dontrun{
code(configuration_object)
}

}
