% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{configuration-class}
\alias{configuration-class}
\title{configuration-class}
\description{
An S4 class to represent a minid configuration file.
}
\section{Slots}{

\describe{
\item{\code{server}}{a character string containing the minid server URL
(e.g. http://minid.bd2k.org/minid)}

\item{\code{user}}{a character string containing the user name (e.g. "Jane Smith")}

\item{\code{email}}{email address for minid user}

\item{\code{orcid}}{the user's ORCID - see https://orcid.org/ (e.g.
"0000-0003-2898-9044")}

\item{\code{code}}{the unique code sent when email address was validated}
}}

\section{see also}{

\code{?\link{configuration}}
}

