% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{configuration}
\alias{configuration}
\title{Create an instance of the configuration class}
\usage{
configuration(server = vector(mode = "character"), user = vector(mode =
  "character"), email = vector(mode = "character"), orcid = vector(mode =
  "character"), code = vector(mode = "character"))
}
\arguments{
\item{server}{a character string containing the minid server URL
(e.g. http://minid.bd2k.org/minid)}

\item{user}{a character string containing the user name (e.g. "Jane Smith")}

\item{email}{email address for minid user}

\item{orcid}{the user's ORCID - see https://orcid.org/ (e.g.
"0000-0003-2898-9044")}

\item{code}{the unique code sent when email address was validated}
}
\value{
an instance of the configuration class
}
\description{
Create an instance of the configuration class
}
\examples{
\dontrun{
example_configuration <- configuration(
             server = "http://minid.bd2k.org/minid",
             user = "Jane Smith",
             email = "jsmith@example.com",
             orcid = "0000-0003-2898-9044",
             code = "XXXX-XXX-XXXXX-XXXX-XXXX"
)
}

}
