% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{content_key<-}
\alias{content_key<-}
\title{content_key}
\usage{
content_key(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new content_key string}
}
\value{
The minid object with specified content_key
}
\description{
Set the minid content_key
}
\examples{
\dontrun{
a_hash <-"6484968f81afac84857d02b573b0d589fb2f9582a2b920572830dc5781e0a53c"
content_key(minid_object) <- a_hash
}

}
