% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{content_key}
\alias{content_key}
\alias{content_key,minid-method}
\title{content_key}
\usage{
content_key(x)

\S4method{content_key}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid content_key slot
}
\description{
Get the minid content_key slot
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the content_key slot
}}

\examples{
data("demo", package = "minidtools")
titles(demo)

}
