% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{created<-}
\alias{created<-}
\title{created}
\usage{
created(x) <- "created string"
}
\arguments{
\item{x}{minid object}
}
\value{
The minid object with specified created string
}
\description{
Set the minid created date
}
\examples{
\dontrun{
created(minid_object) <- "Mon, 19 Mar 2018 17:43:57 GMT"
}

}
