% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{creator<-}
\alias{creator<-}
\title{creator}
\usage{
creator(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new creator string}
}
\value{
The minid object with specified creator string
}
\description{
Set the minid creator
}
\examples{
\dontrun{
creator(minid_object) <- "mdarcy (<id string>)"
}

}
