% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{email<-}
\alias{email<-}
\title{email}
\usage{
email(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new user email string}
}
\value{
The configuration object with specified user email string
}
\description{
Set the minid user email in configuration object
}
\examples{
\dontrun{
email(configuration_object) <- "jsmith@example.com"
}

}
