% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/configuration-getter-methods.R, R/configuration-setter-methods.R
\docType{methods}
\name{email}
\alias{email}
\alias{email,configuration-method}
\alias{email<-,configuration-method}
\title{email}
\usage{
email(x)

\S4method{email}{configuration}(x)

\S4method{email}{configuration}(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new email string}
}
\value{
The email
}
\description{
Get the email from the configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{configuration}: getter method for the email slot

\item \code{configuration}: setter method for the email slot
}}

\examples{
\dontrun{
email(configuration_object)
}

}
