% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/list-resolution-methods.R
\docType{methods}
\name{get_title}
\alias{get_title}
\alias{get_title,minid-method}
\alias{get_title,character-method}
\title{get_title}
\usage{
get_title(minid, n = 1, server = "http://minid.bd2k.org/minid")

\S4method{get_title}{minid}(minid, n = 1, server = NULL)

\S4method{get_title}{character}(minid, n = 1,
  server = "http://minid.bd2k.org/minid")
}
\arguments{
\item{minid}{minid object or identifier to resolve}

\item{n}{index of obsoleted_by minid}

\item{server}{server string}
}
\value{
the nth entry of title minid field.
}
\description{
Get nth entry of title minid field.
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: return the nth title from a minid object

\item \code{character}: do lookup() on minid string and return nth title
}}

\examples{
\dontrun{
get_title(minid = "minid:b94t3q", n = 1, server = server(configuration))
}

}
