% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minid.R
\name{getCreator}
\alias{getCreator}
\title{Query S4 minid object metadata for minid creator}
\usage{
getCreator(obj)
}
\value{
The output is a character string
}
\description{
\code{getCreator} returns the creator from the minid metadata.
}
\details{
This is a generic function: methods can be defined for it.
}
\examples{
demo.id <- "ark:/99999/fk4mp58v5t"
demo.minid <- minid(demo.id)
getCreator(demo.minid)

}

