% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{identifier<-}
\alias{identifier<-}
\title{identifier}
\usage{
identifier(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new identifier string}
}
\value{
The minid object with specified identifier string
}
\description{
Set the minid identifier
}
\examples{
\dontrun{
identifier(minid_object) <- "ark:/57799/b9j69h"
}

}
