% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R,
%   R/minid-setter-methods.R
\docType{methods}
\name{identifier}
\alias{identifier}
\alias{identifier,minid-method}
\alias{identifier<-,minid-method}
\title{identifier}
\usage{
identifier(x)

\S4method{identifier}{minid}(x)

\S4method{identifier}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new identifier string}
}
\value{
The minid identifier string
}
\description{
Get the minid identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the identifier slot

\item \code{minid}: setter method for the identifier slot
}}

\examples{
\dontrun{
identifier(minid_object)
}

}
