% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{load_configuration}
\alias{load_configuration}
\title{Load minid configuration object from config file}
\usage{
load_configuration(config_file = "~/.minid/minid-config.json",
  section = "default")
}
\arguments{
\item{config_file}{Path to the configuration file (default
"~/.minid/minid-config.json")}

\item{section}{if multiple sections in the config file, which one to use
(default "default")}
}
\value{
An object of type "configuration"
}
\description{
Before using the minid API to mint a new minid, you first need to validate
your email address (using register_user()). Following registration, a unique
code will be sent to your email address. You must present this code along
with your email address when accessing the API.
}
\details{
As a convenience you can specify this information in a minid configuration
file (default location is ~/.minid/minid-config.json).

\code{load_configuration()} loads the json-formatted config file and returns
a configuration object.
}
\examples{
\dontrun{

configuration <-
  load_configuration(config_file = "~/.minid/minid-config.json",
                     section = "default")
}

}
