% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/list-resolution-methods.R
\docType{methods}
\name{location}
\alias{location}
\alias{location,minid-method}
\alias{location,character-method}
\alias{title,character-method}
\title{location}
\usage{
location(minid, n = 1, server = "http://minid.bd2k.org/minid")

\S4method{location}{minid}(minid, n = 1, server = NULL)

\S4method{location}{character}(minid, n = 1,
  server = "http://minid.bd2k.org/minid")

\S4method{title}{character}(minid, n = 1,
  server = "http://minid.bd2k.org/minid")
}
\arguments{
\item{minid}{minid object or identifier to resolve}

\item{n}{index of desired location (default 1)}

\item{server}{server string}
}
\value{
The nth location of the identified digital object, and warning if
  minid is obsolete
}
\description{
Get the nth location of an object with given minid
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: return the nth location from a minid object

\item \code{character}: do lookup() on minid string and return nth location

\item \code{character}: do lookup() on minid string and return nth location
}}

\examples{
\dontrun{
location(minid = "minid:b94t3q", n = 1, server = server(configuration))
}

}
