% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lookup}
\alias{lookup}
\title{lookup}
\usage{
lookup(x)
}
\arguments{
\item{x}{minid idenentifier (e.g. ark:/57799/b9j69h)}
}
\value{
A minid object with filled slots
}
\description{
Get a minid by identifier
}
\examples{
\dontrun{
my_minid <- lookup("ark:/57799/b9j69h")
}

}
