% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minid.R
\docType{package}
\name{minid-package}
\alias{minid}
\alias{minid-package}
\title{minid}
\description{
A package for working with BDDS minids.
}
\details{
The \pkg{minid} package curretly includes functions resolving a minid and accessing the
associated metadata.

The \pkg{minid} package does not yet support minting new minids
}

