% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minid.R
\name{minid}
\alias{minid}
\title{minid}
\usage{
minid(id)
}
\arguments{
\item{id}{minid to resolve.}
}
\value{
resolves the  an instance of the minid class
}
\description{
The constructor function for making minid objects
}
\examples{
demo.id <- "ark:/99999/fk4mp58v5t"
demo.minid <- minid(demo.id)
}

