% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{minid_example}
\alias{minid_example}
\title{Get path to load minidtools examples}
\usage{
minid_example(filename)
}
\arguments{
\item{filename}{Name of file}
}
\description{
\pkg{minidtools} comes bundled with sample files in its \code{inst/extdata}
diretory. This function makes them easier to access. Based on
\code{readr::readr_example()}
}
\examples{
minid_example(filename = "demo.rda")

}
