% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minting-methods.R
\docType{methods}
\name{mint}
\alias{mint}
\alias{mint,minid,configuration-method}
\alias{mint,list,list-method}
\alias{mint,minid,list-method}
\alias{mint,list,configuration-method}
\alias{mint,character,configuration-method}
\title{mint}
\usage{
mint(minid, configuration, test = TRUE)

\S4method{mint}{minid,configuration}(minid, configuration, test = TRUE)

\S4method{mint}{list,list}(minid, configuration, test = TRUE)

\S4method{mint}{minid,list}(minid, configuration, test = TRUE)

\S4method{mint}{list,configuration}(minid, configuration, test = TRUE)

\S4method{mint}{character,configuration}(minid, configuration, test = TRUE)
}
\arguments{
\item{minid}{minid object or list of fields or file path}

\item{configuration}{configuration object or list of server parameters}

\item{test}{creates names in a test namespace that is removed periodically
default = TRUE}
}
\value{
a newly minted minid
}
\description{
mint a new minid for a digital object
}
\section{Methods (by class)}{
\itemize{
\item \code{minid = minid,configuration = configuration}: mint a new minid from a minid object

\item \code{minid = list,configuration = list}: mint a new minid from two named lists

\item \code{minid = minid,configuration = list}: mint a new minid from a minid and named list

\item \code{minid = list,configuration = configuration}: mint a new minid from a named list and configuration object

\item \code{minid = character,configuration = configuration}: mint a new minid from a file path and configuration object
}}

\examples{
\dontrun{
# load config
config <- load_configuration()

# mint minid from file path
mint(minid = "./a_file.RDa", configuration = config, test = TRUE)

# or manually set up config
config <- configuration()
user(config) <- "Jane Example"
email(config) <- "jexample@example.com"
orcid(config) <- "0000-0000-0000-0000" # see http://orch.id
code(config) <- "00000000-0000-0000-0000-000000000000" # see register()

# set up minid object
new_minid <- minid()
checksum_function(new_minid) <- "sha256"
checksum(new_minid) <- digest::digest("./somefile.RDa",
                                      file = TRUE,
                                      algo = "sha256")
titles(new_minid) <- list(title = "An Example Object Title")
locations(new_minid) <- list(link = "http://example.com/somefile.RDa") # optional

# mint the new minid
minted <- mint(minid = new_minid, configuration = config, test = TRUE)

# OR use lists:
computed_checksum <- digest::digest(path = "./somefile.RDa",
                                    file = TRUE,
                                    algo = "sha256")
new_minid = list(
  checksum_function = "sha256",
  checksum = computed_checksum,
  titles = "An Example Object Title",
  locations = "http://example.com/somefile.RDa"
  )

configuration = list(
  server = "http://minid.bd2k.org/minid",
  user(config) <- "Jane Example"
  email(config) <- "jexample@example.com"
  orcid(config) <- "0000-0000-0000-0000", # see http://orch.id
  code(config) <- "00000000-0000-0000-0000-000000000000" # see register()
  )

minted <- mint(minid = new_minid, configuration = config, test = TRUE)
}

}
