% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/list-resolution-methods.R
\docType{methods}
\name{newer}
\alias{newer}
\alias{newer,minid-method}
\alias{newer,character-method}
\title{newer}
\usage{
newer(minid, n = 1, server = "http://minid.bd2k.org/minid")

\S4method{newer}{minid}(minid, n = 1, server = NULL)

\S4method{newer}{character}(minid, n = 1,
  server = "http://minid.bd2k.org/minid")
}
\arguments{
\item{minid}{minid object or identifier to resolve}

\item{n}{index of obsoleted_by minid}

\item{server}{server string}
}
\value{
if minid status = TOMBSTONE, the nth entry of the minid obsoleted_by
        field. If minid status = ACTIVE, the minid's identifier.
}
\description{
Get nth entry of obsoleted_by minid field. If minid status = ACTIVE, returns
minid identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: return the nth entry of obsoleted_by minid field. If minid
status = ACTIVE, returns minid identifier.

\item \code{character}: do lookup() on minid string and return nth entry of
obsoleted_by minid field. If minid status = ACTIVE, returns minid
identifier
}}

\examples{
\dontrun{
newer(minid = "minid:b94t3q", n = 1, server = server(configuration))
}

}
