% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{register}
\alias{register}
\title{Register a user}
\usage{
register(config = config)
}
\arguments{
\item{config}{a configuration object with server, user and email slots
filled (mandatory) along with the orcid slot (optional)}
}
\value{
An object of type "minid" or an error if the lookup fails
}
\description{
Before using the API to mint or edit a minid, you first need to validate
your email address. When you register with register(), the server will send
a private code to the email address provided. This code must be supplied for
registering or editing minids. For convienece, this code can
be added to the config object and saved for future sessions.
}
\examples{
\dontrun{

config <- configuration()
user(config) <- "Jane Example"
email(config) <- "jexample@example.com"
orcid(config) <- "0000-0000-0000-0000" # see https://orcid.org/
register(config)

# then get code from email
code(config) <- "0000000-0000-0000-0000-000000000000"
save_configuration(config)

}

}
