% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{save_configuration}
\alias{save_configuration}
\title{Save minid configuration object to config file}
\usage{
save_configuration(config, config_path = "~/.minid/minid-config.json",
  python_style = FALSE, overwrite = FALSE)
}
\arguments{
\item{config}{an object of type configuration}

\item{config_path}{Path to the configuration file (default
"~/.minid/minid-config.json")}

\item{python_style}{write python-style config file (default FALSE)}

\item{overwrite}{overwrite existing configuration file? (default FALSE)}
}
\description{
Before using the minid API to mint a new minid, you first need to validate
your email address (using register_user()). Following registration, a unique
code will be sent to your email address. You must present this code along
with your email address when accessing the API.
}
\details{
As a convenience you can specify this information in a minid configuration
file (default location is ~/.minid/minid-config.json). This configuration
file can be either JSON (default) or the format descrived for the python
minid client package \url{https://github.com/fair-research/minid}.

\code{save_configuration(config)} saves the configuration object config as
a json- or python-style config file at \code{~/.minid/minid-config.json}.
}
\examples{
\dontrun{

config <- configuration(
   server = "http://minid.bd2k.org/minid/",
   user = "Jane Smith",
   email = "jsmith@example.com",
   orcid = "0000-0003-2898-9044",
   code = "XXXX-XXX-XXXXX-XXXX-XXXX"
 )

  save_configuration(config, config_path = "~/.minid/minid-config.json")
}

}
