% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{server<-}
\alias{server<-}
\title{server}
\usage{
server(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new server URL}
}
\value{
The configuration object with specified server URL
}
\description{
Set the minid server URL in configuration object
}
\examples{
\dontrun{
server(configuration_object) <- "http://minid.bd2k.org/minid"
}

}
