% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R,
%   R/setter-methods.R
\docType{methods}
\name{server}
\alias{server}
\alias{server,configuration-method}
\alias{server<-,configuration-method}
\title{server}
\usage{
server(x)

\S4method{server}{configuration}(x)

\S4method{server}{configuration}(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new server URL}
}
\value{
The minid server URL
}
\description{
Get the minid server from the configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{configuration}: getter method for the server slot

\item \code{configuration}: setter method for the server slot
}}

\examples{
\dontrun{
server(configuration_object)
}

}
