% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{short_identifier<-}
\alias{short_identifier<-}
\title{short identifier}
\usage{
short_identifier(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new short identifier string}
}
\value{
The minid object with specified short identifier string
}
\description{
Set the minid short identifier
}
\examples{
\dontrun{
short_identifier(minid_object) <- "minid:b9j69h"
}

}
