% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R,
%   R/minid-setter-methods.R
\docType{methods}
\name{short_identifier}
\alias{short_identifier}
\alias{short_identifier,minid-method}
\alias{short_identifier<-,minid-method}
\title{short identifier}
\usage{
short_identifier(x)

\S4method{short_identifier}{minid}(x)

\S4method{short_identifier}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new short_identifier string}
}
\value{
The minid short identifier string
}
\description{
Get the minid short identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the short_identifier slot

\item \code{minid}: setter method for the short_identifier slot
}}

\examples{
\dontrun{
short_identifier(minid_object)
}

}
