% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,configuration-method}
\alias{show,configuration-method}
\title{show method for configuration object}
\usage{
\S4method{show}{configuration}(object)
}
\arguments{
\item{object}{configuration object}
}
\description{
show method for configuration object
}
\examples{
\dontrun{
show(configuration_object)
configuration_object
}

}
