% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,minid-method}
\alias{show,minid-method}
\title{show method for minid object}
\usage{
\S4method{show}{minid}(object)
}
\arguments{
\item{object}{minid object}
}
\description{
show method for minid object
}
\examples{
data("demo2", package = "minidtools")
show(demo2)

}
