% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{status<-}
\alias{status<-}
\title{status}
\usage{
status(x) <- "status string"
}
\arguments{
\item{x}{minid object}
}
\value{
The minid object with specified status string
}
\description{
Set the minid status
}
\examples{
\dontrun{
status(minid_object) <- "ACTIVE"
}

}
