% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{status}
\alias{status}
\alias{status,minid-method}
\title{Get the minid status}
\usage{
status(x)

\S4method{status}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\description{
Get the minid status
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the status slot
}}

\examples{
data("demo", package = "minidtools")
status(demo)

}
