% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/list-resolution-methods.R
\docType{methods}
\name{title}
\alias{title}
\alias{title,minid-method}
\title{title}
\usage{
title(minid, n = 1, server = "http://minid.bd2k.org/minid")

\S4method{title}{minid}(minid, n = 1, server = NULL)
}
\arguments{
\item{minid}{minid object or identifier to resolve}

\item{n}{index of obsoleted_by minid}

\item{server}{server string}
}
\value{
the nth entry of title minid field.
}
\description{
Get nth entry of title minid field.
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: return the nth title from a minid object
}}

\examples{
\dontrun{
title(minid = "minid:b94t3q", n = 1, server = server(configuration))
}

}
