% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{titles<-}
\alias{titles<-}
\title{titles}
\usage{
titles(x) <- "titles list"
}
\arguments{
\item{x}{minid object}
}
\value{
The minid object with specified titles list
}
\description{
Set the minid titles
}
\examples{
\dontrun{
title_list <- 
  list(paste0("mdarcy - AGR Data set with identifier-based references to ",
    "data in cloud storage"))
titles(minid_object) <- title_list
}

}
