% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{titles}
\alias{titles}
\alias{titles,minid-method}
\title{Get the minid titles}
\usage{
titles(x)

\S4method{titles}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\description{
Get the minid titles
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the titles slot
}}

\examples{
data("demo", package = "minidtools")
titles(demo)

}
