% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{update}
\alias{update}
\title{Update a minid}
\usage{
update(id, config = config, new_title = "", new_location = "",
  new_status = "", new_obsoleted_by = "")
}
\arguments{
\item{id}{the identifier string of the minid to update (e.g.
"ark:/57799/b9j69h" or "minid:b9j69h")}

\item{config}{a configuration object (see \code{confguration()})}

\item{new_title}{the new title for the object (a string)}

\item{new_location}{the new location for the object (a string)}

\item{new_status}{minid status. May be "ACTIVE" or "TOMBSTONE"}

\item{new_obsoleted_by}{the minid identifier to replace the obsolete minid}
}
\value{
An object of type "minid" or an error if the lookup fails
}
\description{
Only locations, titles, locations, status, and obsoleted_by can be updated.
}
\details{
NOTE: lookup currently supports minids of the form "ark:/57799/b9j69h" or
"mind:b9j69h"
}
\examples{
\dontrun{

update(id = "ark:/99999/fk4q53tj2m",
       config = config,
       new_title = "A Better Title!",
       new_location="http://example.com/bar.txt",
       new_status = "TOMBSTONE",
       new_obsoleted_by = "ark:/99999/fk48s5xs8m")

}

}
