% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{user<-}
\alias{user<-}
\title{user}
\usage{
user(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new user string}
}
\value{
The configuration object with specified user string
}
\description{
Set the minid user in configuration object
}
\examples{
\dontrun{
user(configuration_object) <- "Jane Smith"
}

}
