% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/configuration-getter-methods.R, R/configuration-setter-methods.R
\docType{methods}
\name{user}
\alias{user}
\alias{user,configuration-method}
\alias{user<-,configuration-method}
\title{user}
\usage{
user(x)

\S4method{user}{configuration}(x)

\S4method{user}{configuration}(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new user string}
}
\value{
The user
}
\description{
Get the user from the configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{configuration}: getter method for the user slot

\item \code{configuration}: setter method for the user slot
}}

\examples{
\dontrun{
user(configuration_object)
}

}
