% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-methods.R
\docType{methods}
\name{as.list,configuration-method}
\alias{as.list,configuration-method}
\title{as.list}
\usage{
\S4method{as.list}{configuration}(x)
}
\arguments{
\item{x}{configuration object}
}
\value{
a named list
}
\description{
Coerce configuration object to list
}
\examples{
\dontrun{
as.list(configuration_object)
}

}
