% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-methods.R
\docType{methods}
\name{as.list,minid-method}
\alias{as.list,minid-method}
\title{as.list}
\usage{
\S4method{as.list}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
a named list
}
\description{
Coerce minid object to list
}
\examples{
\dontrun{
as.list(minid_object)
}

}
