% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R,
%   R/setter-methods.R
\docType{methods}
\name{checksum}
\alias{checksum}
\alias{checksum,minid-method}
\alias{checksum<-,minid-method}
\title{checksum}
\usage{
checksum(x)

\S4method{checksum}{minid}(x)

\S4method{checksum}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new checksum string}
}
\value{
The minid checksum (as character string)
}
\description{
Get the minid checksum
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the checksum slot

\item \code{minid}: setter method for the checksum slot
}}

\examples{
\dontrun{
checksum(minid_object)
}

}
