% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R
\docType{methods}
\name{checksum_function}
\alias{checksum_function}
\alias{checksum_function,minid-method}
\title{checksum_function}
\usage{
checksum_function(x)

\S4method{checksum_function}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid checksum_function (as character string)
}
\description{
Get the hashing function used to make the minid checksum
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the checksum_function slot
}}

\examples{
\dontrun{
checksum_function(minid_object)
}

}
