% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R,
%   R/minid-setter-methods.R
\docType{methods}
\name{content_key}
\alias{content_key}
\alias{content_key,minid-method}
\alias{content_key<-,minid-method}
\title{content_key}
\usage{
content_key(x)

\S4method{content_key}{minid}(x)

\S4method{content_key}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}
}
\value{
The minid content_key slot
}
\description{
Get the minid content_key slot
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the content_key slot

\item \code{minid}: setter method for the content_key slot
}}

\examples{
\dontrun{
content_key(minid_object)
}

}
