% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{created}
\alias{created}
\alias{created,minid-method}
\title{created}
\usage{
created(x)

\S4method{created}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The date the minid was created (as character string)
}
\description{
Get the minid created date
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the created slot
}}

\examples{
data("demo", package = "minidtools")
created(demo)

}
