% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{creator}
\alias{creator}
\alias{creator,minid-method}
\title{creator}
\usage{
\S4method{creator}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid creator string
}
\description{
Get the minid creator
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the creator slot
}}

\examples{
data("demo", package = "minidtools")
creator(demo)

}
