% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R,
%   R/minid-setter-methods.R
\docType{methods}
\name{creator}
\alias{creator}
\alias{creator,minid-method}
\alias{creator<-,minid-method}
\title{creator}
\usage{
creator(x)

\S4method{creator}{minid}(x)

\S4method{creator}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}
}
\value{
The minid creator string
}
\description{
Get the minid creator
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the creator slot

\item \code{minid}: setter method for the creator slot
}}

\examples{
\dontrun{
creator(minid_object)
}

}
