% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{identifier}
\alias{identifier}
\alias{identifier,minid-method}
\title{Get the minid identifier}
\usage{
identifier(x)

\S4method{identifier}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\description{
Get the minid identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the identifier slot
}}

\examples{
data("demo", package = "minidtools")
identifier(demo)

}
