% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{locations<-}
\alias{locations<-}
\title{locations}
\usage{
locations(x) <- "locations list"
}
\arguments{
\item{x}{minid object}
}
\value{
The minid object with specified locations list
}
\description{
Set the minid locations
}
\examples{
\dontrun{
locations(minid_object) <-
  list("mdarcy - https://nih-commons.s3.amazonaws.com/misc/agr-example.tgz")
}

}
