% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R
\docType{methods}
\name{locations}
\alias{locations}
\alias{locations,minid-method}
\title{locations}
\usage{
\S4method{locations}{minid}(x)
}
\arguments{
\item{x}{minid object}
}
\value{
The minid locations (as list)
}
\description{
Get the minid locations
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the locations slot
}}

\examples{
data("demo2", package = "minidtools")
locations(demo2)

}
