% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{lookup}
\alias{lookup}
\title{Resolve a minid by identifier or file name}
\usage{
lookup(query, configuration = config, hash = "sha256")
}
\arguments{
\item{query}{the minid to look up, specified by file name (e.g.
"file:./some/file.Rda") or identifier (e.g. "ark:/57799/b9j69h", "b9j69h",
or "minid:b9j69h")}

\item{configuration}{the configuration object. Default = config (see also
\code{\link{load_configuration}})}

\item{hash_function}{if performing a lookup on a file, which hash function
to use for querying (default "sha256")}
}
\value{
An object of type "minid" or an error if the lookup fails
}
\description{
Query the minid server specified in the configuration object to retrieve
minid metadata about an identifier or file.
}
\examples{
\dontrun{

my_minid <- lookup(query = "ark:/57799/b9j69h", configuration = config)

my_minid <- lookup(query = "b9j69h", configuration = config)

my_minid <- lookup(query = "minid:b9j69h", configuration = config)

my_minid <- lookup(query = "file:./some/file.Rda",
                   configuration = config,
                   hash = "sha256")

}

}
