% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\name{lookup}
\alias{lookup}
\title{Resolve a minid by identifier or file name}
\usage{
lookup(query, server = "http://minid.bd2k.org/minid", algo = "md5")
}
\arguments{
\item{query}{the minid to look up, specified by identifier (e.g.
"ark:/57799/b9j69h" or "minid:b9j69h") or file name (e.g.
"file:./some/file.RDa")}

\item{server}{the minid resolution server. Can be obtained from
configuration object with server(config). Default =
"http://minid.bd2k.org/minid" (see also \code{\link{load_configuration}})}

\item{algo}{the hashing algorighm to use for file-based minid lookups
(default "md5")}
}
\value{
An object of type "minid" or an error if the lookup fails
}
\description{
Query the minid server specified in the configuration object to retrieve
minid metadata about an identifier or file.
}
\details{
NOTE: lookup currently supports minids of the form "ark:/57799/b9j69h" or
"mind:b9j69h"
}
\examples{
\dontrun{

my_minid <- lookup(query = "ark:/57799/b9j69h", server = server(config))
my_minid <- lookup(query = "minid:b9j69h", server = server(config))
my_minid <- lookup(query = "file:./some/file.RDa", algo = "md5",
                   server = server(config))

}

}
