% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minid.R
\docType{class}
\name{minid-class}
\alias{.minid}
\alias{minid-class}
\title{.minid}
\description{
An S4 class to represent a minid.
}
\section{Slots}{

\describe{
\item{\code{id}}{a character string containing the minid (e.g. ark:/99999/fk4mp58v5t)}

\item{\code{metadata}}{a list of the metadata associated with the minid}
}}
\examples{
getSlots("minid")

}

