% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minid-package.R, R/minid.R
\docType{package}
\name{minid}
\alias{minid}
\alias{minid-package}
\title{minid: A package for working with BDDS minids.}
\usage{
minid(id)
}
\arguments{
\item{id}{minid to resolve.}
}
\value{
resolves the  an instance of the minid class
}
\description{
the \pkg{minid} package allows you to work with minids in R.

The constructor function for making minid objects
}
\details{
\pkg{minid} curretly includes functions resolving a minid and accessing the
associated metadata.

\pkg{minid} does not yet support minting new minids
}
\examples{
demo.id <- "ark:/99999/fk4mp58v5t"
demo.minid <- minid(demo.id)

}
\references{
\url{http://http://minid.bd2k.org/}
}

