% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{minid}
\alias{minid}
\title{Create an instance of the minid class}
\usage{
minid(identifier = NULL, creator = NULL, created = NULL, checksum,
  checksum_function, status = NULL, locations = NULL, titles,
  obsoleted_by = NULL, content_key = NULL)
}
\arguments{
\item{identifier}{a character string containing the minid 
(e.g. ark:/99999/fk4mp58v5t) Note: may be NULL}

\item{creator}{a character string giving the name (and and optionally, ORCID)
of the person who registered the minid (e.g. "mdarcy 
(0000-0003-2280-917X)"). Note: may be NULL}

\item{created}{a character string giving the date and time the minid was 
created (e.g. "Mon, 19 Mar 2018 17:43:57 GMT") Note: may be NULL}

\item{checksum}{the checksum of the identified digital object (e.g. 
"6484968f81afac84857d02b573b0d589fb2f9582a2b920572830dc5781e0a53c")}

\item{checksum_function}{the hashing function used to calculate the checksum 
(e.g. "SHA")}

\item{status}{The minid status (e.g. "ACTIVE" or "MOTHBALLED"). Note: may be
NULL}

\item{locations}{A list of locations hostingthe identified digital object. 
(e.g. "mdarcy - 
https://nih-commons.s3.amazonaws.com/misc/agr-example.tgz"). Note: may be
NULL}

\item{titles}{A list of titles for the identified digital object (e.g.
"mdarcy - AGR Data set with identifier-based references to data in cloud 
storage"}

\item{obsoleted_by}{For minids that have "MOTHBALLED" status, a list of
minids superceding the mothballed minid. Note: may be NULL}

\item{content_key}{a character string, used for the hash of a bdbag. Note:
may be NULL}
}
\value{
an instance of the minid class
}
\description{
Create an instance of the minid class
}
\examples{
test_minid <- minid(
 identifier = "ark:/57799/b9j69h",
 creator = "mdarcy (0000-0003-2280-917X)",
 created = "Mon, 19 Mar 2018 17:43:57 GMT",
 checksum = paste0("6484968f81afac84857d02b573b0d589fb2f9582a2b920572830dc",
                   "5781e0a53c"),
 checksum_function = "SHA",
 status = "ACTIVE",
 locations = list(
   "https://nih-commons.s3.amazonaws.com/misc/agr-example.tgz"),
 titles = list(paste0("AGR Data set with identifier-based references to",
                      "data in cloud storage")),
 obsoleted_by = NULL,
 content_key = NULL
)

}
