% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidtools.R
\docType{package}
\name{minidtools}
\alias{minidtools}
\alias{minidtools-package}
\title{minidtools: An R package for interacting with minids, the MINimal viable 
  IDentifier.}
\description{
The \pkg{minidtools} package provides three categories of tools for working
  with minids:
user registry, querying minids, and creating minids.
}
\section{user registry}{

The foo functions ...
}

\section{querying minids}{

}

\section{creating minids}{

}

\references{
\url{http://http://minid.bd2k.org/}
}
