% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{obsoleted_by<-}
\alias{obsoleted_by<-}
\title{obsoleted_by}
\usage{
obsoleted_by(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new obsoleted_by list}
}
\value{
The minid object with specified obsoleted_by list
}
\description{
Set the minid obsoleted_by
}
\examples{
\dontrun{
obsoleted_by(minid_object) <- list("ark:/57799/b9j69h")
}

}
