% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R,
%   R/setter-methods.R
\docType{methods}
\name{obsoleted_by}
\alias{obsoleted_by}
\alias{obsoleted_by,minid-method}
\alias{obsoleted_by<-,minid-method}
\title{obsoleted_by}
\usage{
obsoleted_by(x)

\S4method{obsoleted_by}{minid}(x)

\S4method{obsoleted_by}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new obsloleted_by list}
}
\value{
The minid obsoleted_by slot (as list)
}
\description{
Get the minid obsoleted_by slot
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the obsoleted_by slot

\item \code{minid}: setter method for the obsoleted_by slot
}}

\examples{
\dontrun{
obsoleted_by(minid_object)
}

}
