% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{orcid<-}
\alias{orcid<-}
\title{orcid}
\usage{
orcid(x) <- value
}
\arguments{
\item{x}{minid or configuration object}

\item{value}{the new orcid string}
}
\value{
The minid or configuration object with specified orcid string
}
\description{
Set the minid or configuration orcid
}
