% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getter-methods.R,
%   R/setter-methods.R
\docType{methods}
\name{orcid}
\alias{orcid}
\alias{orcid,minid-method}
\alias{orcid,configuration-method}
\alias{orcid<-,minid-method}
\alias{orcid<-,configuration-method}
\title{orcid}
\usage{
orcid(x)

\S4method{orcid}{minid}(x)

\S4method{orcid}{configuration}(x)

\S4method{orcid}{minid}(x) <- value

\S4method{orcid}{configuration}(x) <- value
}
\arguments{
\item{x}{minid or configuration object}

\item{value}{the new orcid string}
}
\value{
The minid or configuration orcid string
}
\description{
Get the minid or configuration orcid
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the orcid slot

\item \code{configuration}: getter method for the orcid slot

\item \code{minid}: setter method for the orcid slot

\item \code{configuration}: setter method for the orcid slot
}}

\examples{
\dontrun{
orcid(configuration_object) <- "0000-0003-2898-9044"
}
}
