% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/configuration-getter-methods.R, R/configuration-setter-methods.R
\docType{methods}
\name{orcid}
\alias{orcid}
\alias{orcid,configuration-method}
\alias{orcid<-,minid-method}
\title{orcid}
\usage{
orcid(x)

\S4method{orcid}{configuration}(x)

\S4method{orcid}{minid}(x) <- value
}
\arguments{
\item{x}{configuration object}

\item{value}{the new orcid string}
}
\value{
The orcid
}
\description{
Get the orcid from the configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{configuration}: getter method for the orcid slot

\item \code{minid}: setter method for the orcid slot
}}

\examples{
\dontrun{
orcid(configuration_object)
}

}
