% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{titles<-}
\alias{titles<-}
\title{titles}
\usage{
titles(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new titles list}
}
\value{
The minid object with specified titles list
}
\description{
Set the minid titles
}
\examples{
\dontrun{
title_list <-
  list(list(created = "Mon, 19 Mar 2018 17:43:57 GMT",
    creator = "mdarcy",
    title = paste0("AGR Data set with identifier-based references to",
              "data in cloud storage")))
titles(minid_object) <- title_list
}

}
