% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/minid-getter-methods.R,
%   R/minid-setter-methods.R
\docType{methods}
\name{titles}
\alias{titles}
\alias{titles,minid-method}
\alias{titles<-,minid-method}
\title{titles}
\usage{
titles(x)

\S4method{titles}{minid}(x)

\S4method{titles}{minid}(x) <- value
}
\arguments{
\item{x}{minid object}

\item{value}{the new titles list}
}
\value{
The minid titles (as list)
}
\description{
Get the minid titles
}
\section{Methods (by class)}{
\itemize{
\item \code{minid}: getter method for the titles slot

\item \code{minid}: setter method for the titles slot
}}

\examples{
\dontrun{
titles(minid_object)
}

}
